+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Improvements
 - 样式布局优化
+- 新增单标签页功能
 
 ### fixed
 - 修复bjui-slidebar tooltip 插件销毁错误问题