+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Improvements
 - left-menu侧边栏样式新增手风琴模式
 - 删除部分jQuery废弃函数
+- 修改默认autosize版本为5，兼容1.x版本,需要替换原 **jquery.autosize.js** 为 **autosize.min.js**
 
 -------------------------------------------